Added version property for the endpoint path Removed all v2 specific
