+++ OrdermentumSDK/Routers/AdminRouter.swift
@@ -30,6 +30,13 @@ public enum AdminRouter: URLRequestConvertible {
         }
     }
 
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -68,6 +75,6 @@ public enum AdminRouter: URLRequestConvertible {
 
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try Client.instance.urlRequest(path: path, method: method, parameters: parameters, body: body)
+        return try Client.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body)
     }
 }
