+++ OrdermentumSDK/Routers/AppVersionRouter.swift
@@ -21,6 +21,13 @@ public enum AppVersionRouter: URLRequestConvertible {
         }
     }
     
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -47,6 +54,6 @@ public enum AppVersionRouter: URLRequestConvertible {
     
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try StaticClient.instance.urlRequest(path: path, method: method, parameters: parameters, body: body)
+        return try StaticClient.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body)
     }
 }
