+++ OrdermentumSDK/Routers/DeviceRouter.swift
@@ -24,6 +24,13 @@ public enum DeviceRouter: URLRequestConvertible {
         }
     }
     
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -54,6 +61,6 @@ public enum DeviceRouter: URLRequestConvertible {
     
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try Client.instance.urlRequest(path: path, method: method, parameters: parameters, body: body)
+        return try Client.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body)
     }
 }
