+++ OrdermentumSDK/Routers/ExperimentsRouter.swift
@@ -27,6 +27,13 @@ public enum ExperimentsRouter: URLRequestConvertible {
         }
     }
     
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -61,6 +68,6 @@ public enum ExperimentsRouter: URLRequestConvertible {
     
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try FlagsClient.instance.urlRequest(path: path, method: method, parameters: parameters, body: body)
+        return try FlagsClient.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body)
     }
 }
