+++ OrdermentumSDK/Routers/FindSupplierRouter.swift
@@ -27,6 +27,13 @@ public enum FindSupplierRouter: URLRequestConvertible {
         }
     }
     
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -63,6 +70,6 @@ public enum FindSupplierRouter: URLRequestConvertible {
     
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try Client.instance.urlRequest(path: path, method: method, parameters: parameters, body: body)
+        return try Client.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body)
     }
 }
