+++ OrdermentumSDK/Routers/FlagsRouter.swift
@@ -24,6 +24,13 @@ public enum FlagsRouter: URLRequestConvertible {
         }
     }
     
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -63,6 +70,6 @@ public enum FlagsRouter: URLRequestConvertible {
     
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try FlagsClient.instance.urlRequest(path: path, method: method, parameters: parameters, body: body)
+        return try FlagsClient.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body)
     }
 }
