+++ OrdermentumSDK/Routers/InvoiceRouter.swift
@@ -36,6 +36,13 @@ public enum InvoiceRouter: URLRequestConvertible {
         }
     }
     
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -88,6 +95,6 @@ public enum InvoiceRouter: URLRequestConvertible {
     
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try Client.instance.urlRequest(path: path, method: method, parameters: parameters, body: body, timeout: timeout)
+        return try Client.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body, timeout: timeout)
     }
 }
