+++ OrdermentumSDK/Routers/MarketplaceRouter.swift
@@ -30,6 +30,13 @@ public enum MarketplaceRouter: URLRequestConvertible {
         }
     }
     
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -66,6 +73,6 @@ public enum MarketplaceRouter: URLRequestConvertible {
     
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try Client.instance.urlRequest(path: path, method: method, parameters: parameters, body: body)
+        return try Client.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body)
     }
 }
