+++ OrdermentumSDK/Routers/OrdersRouter.swift
@@ -66,6 +66,13 @@ public enum OrdersRouter: URLRequestConvertible {
         }
     }
 
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -158,6 +165,6 @@ public enum OrdersRouter: URLRequestConvertible {
 
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try Client.instance.urlRequest(path: path, method: method, parameters: parameters, body: body, timeout: timeout)
+        return try Client.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body, timeout: timeout)
     }
 }
