+++ OrdermentumSDK/Routers/ProjectionsRouter.swift
@@ -33,6 +33,13 @@ public enum ProjectionsRouter: URLRequestConvertible {
         }
     }
     
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -75,6 +82,6 @@ public enum ProjectionsRouter: URLRequestConvertible {
     
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try Client.instance.urlRequest(path: path, method: method, parameters: parameters, body: body)
+        return try Client.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body)
     }
 }
