+++ OrdermentumSDK/Routers/SocialRouter.swift
@@ -24,6 +24,13 @@ public enum SocialRouter: URLRequestConvertible {
         }
     }
 
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -54,6 +61,6 @@ public enum SocialRouter: URLRequestConvertible {
 
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try AuthClient.instance.urlRequest(path: path, method: method, parameters: parameters, body: body)
+        return try AuthClient.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body)
     }
 }
