+++ OrdermentumSDK/Routers/SupportRouter.swift
@@ -24,6 +24,13 @@ public enum SupportRouter: URLRequestConvertible {
         }
     }
 
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -54,6 +61,6 @@ public enum SupportRouter: URLRequestConvertible {
 
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try SupportClient.instance.urlRequest(path: path, method: method, parameters: parameters, body: body)
+        return try SupportClient.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body)
     }
 }
