+++ OrdermentumSDK/Routers/VenueNotificationsRouter.swift
@@ -22,6 +22,13 @@ public enum VenueNotificationsRouter: URLRequestConvertible {
         }
     }
 
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -54,6 +61,6 @@ public enum VenueNotificationsRouter: URLRequestConvertible {
 
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try NotificationsClient.instance.urlRequest(path: path, method: method, parameters: parameters, body: body)
+        return try NotificationsClient.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body)
     }
 }
