+++ OrdermentumSDK/Routers/VenueRouter.swift
@@ -72,6 +72,13 @@ public enum VenueRouter: URLRequestConvertible {
         }
     }
 
+    var version: String {
+        switch self {
+        default:
+            return "v1/"
+        }
+    }
+    
     //Paths
     var path: String {
         switch self {
@@ -160,6 +167,6 @@ public enum VenueRouter: URLRequestConvertible {
 
     //Builder
     public func asURLRequest() throws -> URLRequest {
-        return try Client.instance.urlRequest(path: path, method: method, parameters: parameters, body: body)
+        return try Client.instance.urlRequest(path: version+path, method: method, parameters: parameters, body: body)
     }
 }
