--- OrdermentumSDK/Services/ProfileServiceV2.swift
@@ -1,27 +1,0 @@
-//
-//  ProfileServiceV2.swift
-//  OrdermentumSDK
-//
-//  Created by Mark Kenneth Bayona on 11/3/20.
-//
-
-import Foundation
-import Alamofire
-
-public class ProfileServiceV2 {
-    public init() { }
-
-    /**
-     * Fetch the public fields of the signup profile (main purpose for now is to check for autoconnect)
-     * Returns a SupplierSignupProfile
-     */
-    public func checkProfile(code: String, supplierId: String, completion: @escaping (Bool, SupplierSignupProfile?, ErrorResponse?) -> Void) {
-        //Build Route
-        let route = ProfileRouterV2.checkProfile(code, supplierId) as URLRequestConvertible
-
-        //Call API
-        Service<SupplierSignupProfile, ErrorResponse>().request(route: route) { (result, responseObject, errorObject) in
-            completion(result, responseObject, errorObject)
-        }
-    }
-}
